pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "wikisrv"
dynamic = ["version"]
description = "Web server for Markdown knowledge bases"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Typing :: Typed",
]
dependencies = [
    "fastapi>=0.115.0",
    "markdown-it-py>=4.0.0",
    "uvicorn>=0.34.0",
]

[project.scripts]
wikisrv = "wikisrv.cli:main"

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.15.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/wikisrv"]

[tool.hatch.version]
path = "src/wikisrv/__init__.py"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["wikisrv"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
